Fix Polygon.contains at boundary
`Polygon.contains(_:ignoreBoundary:)` with `ignoreBoundary` set to
`false` (the default), was returning `false` for a point at the boundary
that coincides with the bounding box. This is because the optimization
that first checks whether the point is contained by the polygon's
bounding box, which uses `BoundingBox.contains(_:)`, did not consider
a point at the boundary as being contained.

To address this, `BoundingBox.contains(_:)` is extended with a second
argument `ignoreBoundary` defaulting to `true` for backwards
compatibility. The optimization in `Polygon.contains(_:ignoreBoundary:)`
now forwards the `ignoreBoundary` argument, fixing the issue.

Fixes #88