Fix Polygon.contains at boundary of hole
When including the boundary, a point at the boundary of a hole is
considered to be contained by the polygon. Conversely, when not
including the boundary a point at the boundary of a hole is considered
not to be contained by the polygon.

Neither of those were the case as shown by the added test. The fix is
as simple as flipping the `ignoreBoundary` flag when checking whether
the point is within one of the holes.